testharness: use mutable Diagnostics slice for builder collection
When we're loading a spec using builders, any of the builders can generate
diagnostics in any order, depending on user input. Therefore it's more
convenient to have all of them append to a single shared diagnostics
slice than to have each one produce its own and concatenate them together
later, and this method also preserves the call order in the returned
slice.