core: Fix TestContext2Apply_countDecrease
This test's original outcome was invalid because it didn't actually
configure a mock apply function, and so _any_ apply operation would appear
to have destroyed the object it was given. This was visible in the fact
that the configuration contains aws_instance.bar but yet it was not
present in the expected state string.

Now we use the standard testApplyFn, and update the expected output to
include the aws_instance.bar object that is created by a Create change.