lang/langserver: Simple implementation of Formatting
Mainly just to have a proof-of-concept of this doing something, this is a
simple implementation of the Formatting method that just replaces the
whole buffer with the hclwrite result.

In future we should make this instead return surgical TextEdits, but we'll
need more infrastructure in place here before that is possible.