Make sure existing paths are updated during `init`
Without this additional code new paths would be written correctly using a slash, but existings paths would not be updated.

Things would keep working without issues, but it would not fix the use case this PR was created for (using remote operations from Windows to TFE). At least not for existing users.