command/0.12checklist: Report invalid resource and provider alias names
Terraform 0.11 had incomplete validation of resource and provider alias
names, which worked only due to the strange way HIL parsed variable names.
Terraform 0.12 (based on HIL) requires valid identifiers, so we'll let
users know about this before they upgrade. It will be easier for users to
deal with this while still on Terraform 0.11 because it won't run into
parse errors trying to work with the invalid name.