command/0.12checklist: Detect reserved variable names
Although it's not super important to deal with these before upgrading to
0.12, some prior warning of it during checklist will be helpful to those
with complicated, decomposed environments so that they can avoid starting
on the upgrade process and then finding out partway that they have some
module somewhere that needs a breaking change to fix this.

Because it requires a breaking change to the module to resolve, this is
something that may be better dealt with as a separate change before
beginning the 0.12 upgrade, to avoid changing too much at once.