Various bug fixes.
Actually initialize the ProviderMetas map for the Module when calling
configs.NewModules so we don't panic.

Check the NodeValidatableResource implements
GraphNodeAttachProviderMetaConfigs.

When calling terraform.loadProviderSchemas, actually set the
ProviderMeta block.

Add ProviderMeta to state. For some reason? I don't know why.

Fix the ProviderTransformer to use the ProviderConfig.Type instead of
String() output for the provider when looking up the ProviderMeta to
use.