command: Make room for project.Workspace objects
In our existing architecture, workspaces belong to backends and so the
command package only deals with workspace _addresses_. With workspaces
becoming a project concept, we need to make room for preparing and passing
around project.Workspace objects, which requires some renaming to make it
clear when we're dealing with actual workspace objects vs. workspace
addresses.