vendor: A specific version of protoc-gen-go
We already depend on the parent protobuf package anyway, so this doesn't
actually add a new dependency, but it means that if we build protoc-gen-go
in vendor mode then we'll get the version that's vendored here, while
a non-vendor install would already have picked up this same version from
go.mod anyway.

This shouldn't change any behavior in practice; it's just so that
things work as expected when we run in -mod=vendor mode.