Update our test cases to use submodules.
Add a submodule to our test configs so we can check that this actually
works across module boundaries the way it's intended to.

These test configs don't actually do anything yet, our test code needs
to be updated to actually check for them, and I'm not sure how to do
that just yet, but I'll figure that out.